Add a "Loading..." notification.
I did this by creating a "wrapper" library for jquery, where the
wrapper injects some callbacks to set up the loading indicator.

This doesn't seem that horrible, but I wish I could rename the
original 'jquery' library to something like 'jquery_raw' so that
people know not to use it. But alas, it seems jquery does not like it.